WIP: Add departed state for members
Fixes #1994.

This PR adds a fourth state for members, 'Departed'. A departed member
is no longer counted as valid for calculating the total population
needed for a quorum.

When a member is marked `Confirmed`, we record the time from the expired
thread. When the expired thread runs, it checks for members whose
departure time has come; and if it has, it sets their health to
`Departed`, removes them from the departure timer, and shares the
rumor.

This commit sets the departure timer at 72 hours, a particularly
conservative number. If an administrator does nothing to bring member
back to life in that timeframe, it will be marked departed
automatically.

Additionally, when a service group has confirmed members, and new
members are added to the service group, the supervisors will
automatically mark a confirmed member as departed. They do this in
alpha-sort order for the members ID's. This will ensure that we handle
the typical recovery scenario without any administrator intervention.
Hat tip to the clever folks at Consul for this solution.

TODO:

* Have a command line tool in `hab-butterfly` to manually mark/unmark a node as
  Departed

Signed-off-by: Adam Jacob <adam@chef.io>